removed signout from homepage and implemented in the NavBar
